app/requirements.txt
flask==2.0.2
requests==2.26.0
psycopg2-binary==2.9.6
boto3==1.28.1
json2html==1.3.0